+++ packages/frontend/src/providers/DashboardProvider.tsx
@@ -109,7 +109,6 @@ export const DashboardProvider: React.FC<{
                     const overriddenDimensions = applyDimensionOverrides(
                         d.filters,
                         schedulerFilters,
-                        true,
                     );
 
                     return {